refactor: remove deprecated chat module and centralize WebSocket logic
