--- src/chat/__init__.py
@@ -1,5 +1,0 @@
-from src.chat.router import router
-
-__all__ = [
-    "router",
-]
